商品和规格关联完成

+++ hope-winery/src/main/java/com/ruoyi/winery/domain/winery/WineryOrders.java
@@ -80,7 +80,7 @@ public class WineryOrders implements Serializable {
      * 商品基准单价
      */
     @Excel(name = "商品基准单价")
-    private Long goodsPrice;
+    private BigDecimal goodsPrice;
 
     /**
      * 商品数量
